代码/服务器版本/服务器端/requirements.txt
Pillow==8.4.0
pypinyin==0.46.0
requests==2.27.1
timeout_decorator==0.5.0
urllib3==1.26.10
